Fixed bug in FileSpecHelper
When the utility function for the handle, comp, parent_handle is called
in a 'size' context there was a bug:

process (p, isFullPath);
if (isComponent)
{
process (p, parent_handle);
process (p, component);
}
else
{
process (p, handle);
}

isFullPath was not being initialized because process (p, isFullPath)
only uses type information to determine the size.

The underlying reason is: when asked for the min and max size of this
type (FileSpecHelper) we cannot give an accurate number until we know if
a full path is used or not. Therefore, just take the worst case
and use max () and min() of both branches. However, since this is harder
to do, just use the sum (making the memory requirements a bit higher
than they really need to be).