Fix bug in LogAPI for readlink
boost::format ignores the format specifier and uses
the variable type instead.
a char * pointer was passed in as "%p" with the intention of
printing the pointer address, but boost::format instead tried to print
the string. Valgrind flagged this in case of an error readlink, in which
case the buffer was initialized.