modified sysio cflags and MPI detection.
MPI detection occurs early on in configure and will
overide CC, CXX, and CPP if mpicc, mpicxx, and mpicpp
if they are found when --with-mpi= is specified (else abort). This ensures
that the mpi compilers are used for the compile tests in configure. The sysio
flags should be in this file instead of the Makefile.subdir.