added a preallocated object pool for IOFWD tasks.
Instead of dynamically allocating a new task for each request,
grab a preallocated task from the TaskPool and reset the parameters.
Config file param support for adjusting the number of preallocated
tasks in the pool. At the moment, only WriteTasks are supported.
This is a WIP.

Whietspace cleanup on all modified files.