Fix bug in Singleton
Singleton wasn't safe it is was called from within another static
initialization function. The problem was that Singleton contained a
non-pod static scoped_ptr variable, and there was no guarantee about when its
initialization function (setting the scoped_ptr to 0) was ran.

If it ran before the static initialization function that used the
singleton, singleton was used the ptr uninitialized; In addition, at
some point after that the constructor of scoped_ptr would run, resetting
ptr to 0 while the init_flag was set to true; the next user of the
singleton would get a 0 ref...