updated the read and write tasks (thread and SMs) to use the BMI memory manager
This is a WIP. Right now, the bmm is a singelton and this may change. If it does
all of the tasks will need to hold a pointer to the bmm. Additionaly, the read and
write tasks allways allocated their own buffers for non-pipeline transfers. The
tasks must now allocate all buffers through the memory manager.