Added ZoidFSAsync implementation
This async API implementation calls the blocking API; As such, it is not
asynchronous but will block the caller. It can be used for devices that
are so fast in performing I/O that it doesn't make sense to put the
request on an internal threadpool. (For example, /dev/null)