Fixed nasty bug in Configurable
Configurable::configure_if_needed was catching bad_cast to see
if the underlying class supported the Configurable interface.
Unfortunately, it was also calling the configure method of the class
within its try block. As such, it would catch and hide any bad_cast
exception (or derived!) and act as if the underlying object did
not support configuration.