TimerResource fix/enhancement
In some rare cases, cancel could return false *before* the callback was
called. When cancelling the next timer that will fire at exactly the
fire time, the worker thread could grab the timer entry first. It has
the lock while doing this. It then drop the locks before executing the
callback.

If cancel is called concurrently and it doesn't get the lock first, it
will no longer see the timer entry and think it was already executed. In
reality, it is concurrently executing on the worker thread. It returns
false indicating that it couldn't cancel the timer because it already
executed. In reality, it is still executing and cancel could return
before it is done executing.

This violates the rule that cancel will either cancel the callback or
fail to do so (because it was too late) in which case it will wait until
the callback completes.  After cancel returns it guarantees that the
callback associated with that timer entry already completed or was
called with 'CANCELLED'.