code cleanup for the request merger / range scheduler
All range sets must now derive from BaseRangeSet. Updated
the existing range sets to derive from this new base class. Updated
the range scheduler functions to only use pointers to the base class.
This removed a bulk of the #defines in that code. Most of this
work is in prep for making the request merging code configurable.