Make zoidfsclient return error instead of exiting
zoidfsclient.c was calling exit(1) on all BMI errors, breaking things
like fuse.

Lightly tested fix (works with fuse). Only fixes the blocking code path
in the client.

With this path, fuse returns 'function not implemented' (the translation
of ZFSERR_OTHER) if the forwarding server is down. If the/a server comes
back up, it reconnects and works again.