updated BMIMemoryManager usage in the tasks, SMs, and frontend
Removed all refs to the BMIMemoryManager, BMIAddr, and BMIBuffer
from the iofwd/tasks and iofwd/tasksm/SMs. Also removed BMI refs
in the Request base classes. Instead, all of the transport specific
memory allocation is done in the iofwd/frontend Requests. The
IOFWDFrontend also starts and shuts down the mem manager isntead of
the DefRequestHandler. The memory manager config was moved into the
frontend section of the config file (removed from the request handler section).
A general memory manager interface was developed (IOFWDMemoryManager),
BMIMemoryManager implements that interface, and the memory management
code is in iofwdutil/mm .