Optimize getSource() call
Before this patch, getSource() defaulted to getSource("default") meaning
the getSource() would do the lookup in the list of already created
sources.

This patch adds an optimization for getSource() so that no lookup is
required, making it faster to just use getSource() for temporary
logging/tracing as the result of getSource() does no longer need to be
cached due to performance reasons.