Split GenericTransform factory into encode/decode
There are now 2 GenericTransform factories: one for encoding, one for
decoding. Something encoded with key 'LZF' can be decoded with the
decode factory and key 'LZF'.

Updated transforms.

Updated unit test so that it automatically determines which transforms
can be tested (the ones for which encode and decode exist).

-- NOTE: Breaks unit test since the default 'ZLIB' does not support
reset...