adjusted the ThreadPool users to use the updated submitWork() interface
Adjusted the SMManager to use submitWork and cached the TP ref. Updated
the RequestScheduler to use submitWork and cached the TP ref. Updated
the DefRequestHandler to cache the TP ref. Updated the ZoidFSAsync API
to use submitWork and cached the TP ref. Added arg wrappers to
the ZoidFSAsync api to reduce the number of binds and allow boost
to optimize the bind of one arg to a static member function.