bug fix: programs using the local lib (no forwarding) crashed
The zoidfs handler setup expected that the dispatchers to use
were previously setup. Works in the server because setup was done
at a higher level. Failed in the local test lib because the
handlers to be used were never initialized. For now, use
preprocessor conditional to enable the local handler code.