updated client to support more advanced communication patterns.
Added client code to allow swapping out the peer address with
another user supplied address. Updated the router to support this new
mode. Added a file block alignment router mode that distributes
blocks to servers that "own" or "manage" file regions that the
blocks are a subset of. Router lib now uses the weak symbol
interface for the zoidfs lib to overload the default behavior
of the zoidfs client.