Limit tags used by client to [0,65536[ range.
Switched to TLS and give each thread its own tag.

Note: this is not an ideal solution, since the code will assert once
65536 threads have made a zoidfs request. Solution is going with OpenPA;
However, the client currently assumes get_tag() always returns the same
value for the same thread, so before we can switch to true dynamic tag
values that code needs to be fixed.