Add backtrace support to ZException
if --enable-exception-backtrace is passed to configure, ZTHROW will
include a stack trace in the exception object. If the exception is
shown using diagnostic_information, the symbol names belonging to the
backtrace will be looked up, demangled and displayed.

While the actual function names are only constructed if the exception
object is actually converted to a string, the exception object will
always contain an array containing the function pointers and thus adds
time and memory overhead (+- 36*sizeof(void *) bytes).