Fix signal handling
In the move to the service system, the order in which threads were
created changed. The server handles signals by redirecting everything to
a single thread which waits for signals. To do this, it first disables
all signals so that any thread started from the main thread inherits the
no-signal mask. Due to the thread reordering, some threads did not have
the no-signal mask; Because of this, sending ctrl-c/usr1 to the server
had a 50/50 change of just killing the server instead of doing a clean
shutdown.

The run-local-test.py for the service-migration commit didn't catch this
(as sometimes it would work, sometimes it wouldn't)

Fixed.