updated hint processing in the zoidfsclient
all hint-enabled zoidfs calls can now handle a hint
passed back from the server and account for the
hint size in the recv buffer. After the unexpected
sends and before the server response, all key /
values in the hint are purged. The hints set by the
server are passed back to the caller of zoidfs.