Add user error message support to ZException
Added addUserErrorMessage () and getUserErrorMessage () functions in
ZException.hh, which can be used to store a user displayable message in
the exception object.

For exceptions that make it to iofwd.cpp main, we can only show the user
error message (and dump the rest as developer info). Exceptions that get
there and don't have the user message should be considered programmer
faults (like failed assertions).