zoidfs_create throttle optimization
If concurrent calls to zoidfs_create are made (for the same filename!),
block all subsequent calls until the first one completes.

If the first one completes /without error/ AND /created == 1/, don't try
to do the create again but return the same information (handle), but
with created = 0.

In all other cases, the call is going to be retried (at which point the
process repeats) but with one less caller. Meaning that if 5 concurrent
calls are made, and the first one returns an error, the remaining 4 will
retry and the first of the four will try to create while the rest waits.

NOTE: this relies on posix_simplify_path returning exactly the same
filename; Situations such as hard links, path components using symlinks,
... will not be detected and the optimization will not be triggered.