Add performance counter to monitor write performance.
Five counters are added by this commit, here is the details:
- write_sm: Counter on how many write state machine tasks are created.
  It will be increased by one when a new write request arrives.
- write_sm_livetime: The duration from a specific write request is
  created to the time it is deleted.
- write_sm_recvtime: The time duration spent to receive the data buffer
  from the client in a write request.
- write_sm_waittime: The time duration spent in waiting for CPU in the
  task queue for every write request.
- write_sm_fsiotime: The time duration spent in performing IO requests
  to the backend file system.

This commit added a dumper thread to dump all the counters to std::cout
periodically. You can start this thread by giving a positive number to
"counters { dumpinterval = %% }" in the configure file.

These counters could be enabled separately in the configure file, this
commit also demonstrated how to do that.

Change-Id: Ieaf4d1f6c2ed71e8c17df3ab1f588be133c88a5a