Revert "Fix for page cache issue from Gary"
Previously, we found that when the memory is full of cached pages
(even they are all clean) the performance will drop significantly.
So we clear all cached pages when the file is closed.

However now we found that the problem is in Linux kernel and it has
been fixed in CentOS 6.2. So the previous patch is unnecessary if
we avoid to use SLES 11.

This reverts commit a62a6eea75041e7418f89e8d3dc6798ab67b254a.